html2: Split out tokenizer tokens into their own file
